requirements.txt
crewai==0.51.1
crewai[tools]==0.51.1
google-generativeai==0.8.3
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyGithub==2.8.1
tree-sitter==0.22.3
tree-sitter-python==0.21.0
pylint==3.2.7
bandit==1.7.8
radon==6.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
alembic==1.13.2
celery==5.4.0
redis==5.1.1
structlog==24.4.0
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
PyJWT==2.8.0
cryptography==42.0.0
